Refactoring SwiftExpressionKind
